split: extract WorkspaceSettingsService from AdminService
Moves createWorkspace, updateWorkspaceSettings, updateTheme,
updateEmailBranding, updatePortalLayout to a focused service. AdminService
drops 4 dependencies (themeRepository, portalConfigRepository,
emailBrandingRepository) and shrinks 1145 → 813 lines.
