+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminApiKeysTest.kt
@@ -284,6 +284,9 @@ class AdminApiKeysTest {
             }
             adminRoutes(
                 adminService = buildAdminService(),
+                workspaceSettingsService =
+                    com.kauth.domain.service
+                        .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
