+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminRoutesTest.kt
@@ -607,6 +607,9 @@ class AdminRoutesTest {
         routing {
             adminRoutes(
                 adminService = buildAdminService(),
+                workspaceSettingsService =
+                    com.kauth.domain.service
+                        .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
@@ -688,6 +691,9 @@ class AdminRoutesTest {
         routing {
             adminRoutes(
                 adminService = buildAdminService(),
+                workspaceSettingsService =
+                    com.kauth.domain.service
+                        .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
