+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminWebhooksTest.kt
@@ -287,6 +287,9 @@ class AdminWebhooksTest {
             }
             adminRoutes(
                 adminService = buildAdminService(),
+                workspaceSettingsService =
+                    com.kauth.domain.service
+                        .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
